# docs/api/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/signup:
    post:
      tags:
      - user-sign-up-rest-controller
      summary: Redirect the user to the signUp view
      operationId: signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisteredRequest'
        required: true
      responses:
        "200":
          description: Registration successful
          content:
            application/json: {}
        "400":
          description: Invalid request
...
  /api/verify:
    get:
      tags:
      - user-sign-up-rest-controller
      summary: Check the verification code to create an account
      operationId: verifyUser
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Account verified
          content:
            application/json:
              schema:
                type: object
        "400":
          description: invalid Token
...
components:
  schemas:
...
    RegisteredRequest:
      type: object
      properties:
        email:
          type: string
        username:
          type: string
        password:
          type: string
...